[WIP] Refactor connmanager
